+++ src/main/java/org/c3lang/intellij/psi/impl/C3PsiImplUtils.java
@@ -3,24 +3,29 @@ package org.c3lang.intellij.psi.impl;
 
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.PsiElement;
-import org.c3lang.intellij.psi.C3StructDeclaration;
-import org.c3lang.intellij.psi.C3Types;
+import org.c3lang.intellij.psi.*;
 
 public class C3PsiImplUtils
 {
-    public static String getName(C3StructDeclaration decl)
+    public static String getName(C3TypeDecl element)
     {
-        ASTNode node = decl.getNode().findChildByType(C3Types.TYPE_IDENT);
+        ASTNode node = element.getNode().findChildByType(C3Types.TYPE_IDENT);
         if (node == null) return null;
         return node.getText();
     }
 
-    public static PsiElement setName(C3StructDeclaration decl, String newName)
+
+    public static PsiElement setName(C3TypeDecl decl, String newName)
     {
+        ASTNode nameNode = decl.getNode().findChildByType(C3Types.TYPE_IDENT);;
+        if (nameNode == null) return decl;
+        C3TypeName name = C3ElementFactory.createTypeName(decl.getProject(), newName);
+        decl.getNode().replaceChild(nameNode, name.getNode());
         return decl;
     }
 
-    public static PsiElement getNameIdentifier(C3StructDeclaration decl)
+
+    public static PsiElement getNameIdentifier(C3TypeDecl decl)
     {
         ASTNode node = decl.getNode().findChildByType(C3Types.TYPE_IDENT);
         return node == null ? null : node.getPsi();